requirements.txt
streamlit==1.52.2
pandas==2.3.3
python-dotenv==1.2.1
neo4j==5.28.2
langchain-openai==1.1.7
langchain-community==0.4.1
langchain-experimental==0.4.1
langchain-core==1.2.6
pypdf==6.5.0
langgraph
langchain-classic==1.0.1
langchain-neo4j==0.6.0
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.1
layoutparser==0.3.4
lxml==6.0.2